✨ feat(local-file-mention): implement useLocalFileMention hook for local file search functionality
